Update issue 2351
Fixing a problem with the representation of element validity for
primitives.  Part of this includes implementing a RepeatingTimer object,
so I can check on the state of another Thread at a given interval.

Also making it so that a Form() can run all the way through its target
execution when calling submit().