Update issue 2333
Working on a bug saying that it isn't safe to use QPixmaps outside the
GUI thread.  Presumably this is because changing the color of the text
is, in fact a QPixmap.

It doesn't seem to be breaking anything at the moment, so I'll just
leave it alone for the time being.