Update issue 2333
This is a massive commit because I want to be sure to have a snapshot of
the code that doesn't segfault (or at least, it's harder to reproduce
... still need to test it on my mac.)

The key feature of this commit is that I'm implementing updates to the
icons of the Qt layer with their signal/slot mechanism, which eliminates
the QPixmap issue I've so regularly experienced.

Also included in this commit is a ton of attempts to make the GUI appear
as expected on Linux.  This is unfortunately still an issue. [ background
info on this issue: issue 2420.  I'm able to see the UI as expected on
Mac OS X, but not on Linux. ]