Update issue 2435
Adding a massive test function to the GUI integration test for the
collapsible container.  I'm realizing now that I really need a separate
test file to test the state of Qt widgets ... the integration test
really needs to test the integration of the core with Qt, not handle all
the Qt testing as well.