Update issue 2507
Using a new object in the i18n module to track the current language.
This way, the default language might be set to EN, but we can switch it
later according to user preference.

I suspect that we'll be able to make a very reasonable guess about the
user's preferred language based on the operating system's locale later
on.  This new functionality will allow us to adjust the translation
(which also affects the JSON translation) at application startup.