Fixes issue 2537
Making it possible for logging runs to re-use the message dialog.

This issue was happening because I had included a logging filter on the
LOGGER object that was restricting anything from passing through that
wasn't from the target thread.  Since each model run takes place in a
single thread, I can't just leave the filter in place from a previous
thread if I want to re-use the same logging handler.  So now I clean up
the logging handler and the logging object at the end of the model run.