Update issue 2823
The issue was that the messages dialog was not being added as a handler
until well after the form had been submitted and the operation thread
had started.  The current structure now has the messages dialog
presented to the user (and the messages logging handler) added to the
logging package after the PythonRunner is created and just before the
execution thread is started.