Update issue 2962
Ensuring that I don't inadvertently set a requirement value as part of
validating field requirements.  This was the root of the problem
detailed in the issue.  It was causing extra information to be included
in the hashable configuration object, but only after validation had
completed for that element.

Because lastrun details are loaded near the beginning of the UI's
existence, elements hadn't had a chance to validate, so the UI was
unable to fetch the correct value from the lastrun file due to the
difference in hashable object details,