Fixing issue with merging multiple default dicts
Merging default dictionaries is not as simple as just taking the union
of the two … it can also involve taking the current default state into
account.  If we have a rendered configuration dictionary entry that is
the same as the old default, then it should be converted to the new
default value, whatever that may be.