Fixing an issue with interactivity of textfields
When the user changed text, this caused the text in a text field to
eventually be set pragmatically to the same value as the user had
defined.  This caused the cursor to be cast to the end of the line.
Adding a check for when the user is editing the textfield.