Correcting container element toggling.
When a container GUI is created, we shouldn’t have the container be
responsible for initializing the state of the core application … that
should already be complete BEFORE the GUI is set up!

The core application SHOULD be affected when the GUI layer is altered
by user interaction, however.

Revealed when fixing for Crop Production.