Supporting OGRFieldDropdowns
This change introduces several new features, ultimately in service of
adding a palisades equivalent of the OGRFieldDropdown element.
  * A new abstract TableDropdown class has been added
  * OGRFieldDropdown (subclass of TableDropdown) provides OGR-specific
support for loading field names from a vector.
  * There’s a new shorthand signal, ‘populate_dropdown’, used for
sending the value of a file or text element to the load_columns method
of a TableDropdown.
  * Dropdown elements now support having their options set by the user
arbitrarily, emitting an appropriate signal.
  * The GUI layer now responds by updating its contents when a dropdown
element’s options have been updated.

These changes are required for Finfish Aquaculture.