Refactor of communication and submission error handling
Communicator objects now support kwargs on emit() calls as well as when
the call is registered.  This could be a little tricky
documentation-wise, I admit, since a registered callback might not
all have the same arguments.  Still need to think about that.

Communicator objects may now be assigned names.  This is particularly
useful when debugging which thread made which callback.

Workspace existence checking is now a special case in the form
submission process, but the callback is started like any other callback
in the form submissin process.